test(functions): testes das funções puras como gate de deploy + shared/ (Onda 3)
Os testes do veredito/matching (F2), das faixas do resumo e do stake (item 15)
viviam em scripts ad-hoc fora do repo — ninguém conseguia rodar de novo. Agora
são deno test no CI, gate dos DOIS deploys (needs: test) + feedback em todo PR.

Refactor move-only (zero mudança de lógica) pra tornar o código testável:
- notify-settlement/verdict.ts: norm/hasTeam/aliases/teamNames/fixtureTeamNames
  + computeVerdict/settleAsianHandicap + tipos (extraído de index.ts)
- notify-weekly-summary/tiers.ts: pickTier/buildMessage/thresholds
- telegram-webhook/stake.ts: parseStake + isBareNumber (regex do interceptador)
- shared/format.ts (esc) e shared/links.ts (hmacHex/trackedUrl) — consolidação
  SÓ do que estava duplicado 1:1 (deriva entre cópias quase causou bug; ver
  norm/545ccbb). notify-opportunities e weekly passam a importar.

Testes: verdict (27 casos — matching multi-liga, anti-falso-positivo
Atlético-MG≠GO≠Madrid, decimal intacto, ML/OU/BTTS/AH quarters, conservadorismo)
+ weekly (11 — thresholds e copy por faixa, nunca taxa de acerto) + stake (13 —
"aposta real nunca intercepta"). Mini-assert local: 100% offline.

MAPA: alerta do ops-healthcheck registrado na seção Interno (#11).
